feat: server without DB

+++ server/src/main/java/com/simeon/commands/clientmanager/RemoveAtCommand.java
@@ -1,9 +1,12 @@
 package com.simeon.commands.clientmanager;
 
 import com.simeon.Response;
+import com.simeon.Role;
+import com.simeon.UserInfo;
 import com.simeon.collection.ICollectionManager;
 import com.simeon.commands.Command;
 import com.simeon.element.Organization;
+import com.simeon.exceptions.DeniedModificationException;
 import com.simeon.exceptions.NoSuchParameterException;
 import lombok.NonNull;
 import lombok.extern.java.Log;
@@ -25,7 +28,7 @@ public class RemoveAtCommand extends Command {
     }
 
     @Override
-    public Response execute(@NonNull Map<String, ? extends Serializable> parameters) {
+    public Response execute(@NonNull Map<String, ? extends Serializable> parameters, @NonNull UserInfo userInfo) {
         log.log(Level.INFO, "{0} command command started with {1}", new String[]{name, parameters.toString()});
         int index;
         try {
@@ -39,9 +42,15 @@ public class RemoveAtCommand extends Command {
         }
 
         try {
-            collectionManager.deleteAt(index);
-            return new Response(true,
-                    String.format("The item by index %d has been successfully deleted", index));
+            if (userInfo.getRole().compareTo(Role.ADMIN) >= 0 ||
+                    collectionManager.getAt(index).getUserInfo().getId() == userInfo.getId()) {
+                collectionManager.deleteAt(index);
+                return new Response(true,
+                        String.format("The item by index %d has been successfully deleted", index));
+            }
+            else {
+                return new Response(false, new DeniedModificationException());
+            }
         }
         catch (IndexOutOfBoundsException e) {
             return new Response(false, e);
